exception added

+++ src/users/users.service.ts
@@ -36,6 +36,9 @@ export class UsersService {
 
     async removeUser(id:number){
         const user = await this.repo.findOne({where:{id}, cache:false});
+        if (!user){
+            throw new NotFoundException("user not found");
+        }
         this.repo.remove(user)
     }
 }
